pyproject.toml
[project]
name = "aratrust-eval"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "httpx>=0.28.1",
    "langchain>=1.1.0",
    "langfuse<3.0.0",
    "langgraph>=1.0.4",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "tqdm>=4.67.1",
]
